fixed errors and stuff

+++ botthing.py
@@ -12,19 +12,10 @@ usrdatafile = open(tokenpath, 'r')
 usrdata = usrdatafile.readlines()
 
 
-mushdatafile = open(datapath, 'r')
-mushdata = mushdatafile.readlines()
-mushdatafile.close()
-
-
 # reading data
 __THING__ = b64decode(usrdata[0]).decode('utf-8')
 sheryel = int(b64decode(usrdata[1]).decode('utf-8'))
 awex = int(b64decode(usrdata[2]).decode('utf-8'))
-
-
-mushroomnum = int(mushdata[0])
-
 
 messages = ['uwu', 'x3', ':3', 'rawr']
 real_spellings = ['real', 'rael', 'rail', 'rewal',
@@ -36,59 +27,51 @@ intents = discord.Intents.all()
 
 bot = commands.Bot(command_prefix='$', intents=intents)
 
-@bot.command(name='test')
-async def a0(ctx, *, arg):
-    print('real')
-    await ctx.send(arg)
-
 @bot.event
 async def on_message(message):
-    if message.author.id == sheryel:
+    if not message.author.id == 1115615704071282750:
+        if message.author.id == sheryel:
 
-        # kiss dear awex
-        if(message.content == 'kiss boyfie' or message.content == 'kiss awest' or message.content == 'kiss awext'):
-            await message.channel.send("*kisses* <@" + str(awex) + "> " + messages[randint(0, 3)])
+            # kiss dear awex
+            if(message.content == 'kiss boyfie' or message.content == 'kiss awest' or message.content == 'kiss awext'):
+                await message.channel.send("*kisses* <@" + str(awex) + "> " + messages[randint(0, 3)])
 
-        # i wuv boyfie
-        if(message.content == 'i wuv boyfie' or message.content == 'i wuv awest' or message.content == 'i wuv awext' or message.content == 'wuv awext' or message.content == 'wuv awest' or message.content == 'wuv boyfie'):
-            await message.channel.send("i wuv you <@" + str(awex) + "> " + messages[randint(0, 3)])
+            # i wuv boyfie
+            if(message.content == 'i wuv boyfie' or message.content == 'i wuv awest' or message.content == 'i wuv awext' or message.content == 'wuv awext' or message.content == 'wuv awest' or message.content == 'wuv boyfie'):
+                await message.channel.send("i wuv you <@" + str(awex) + "> " + messages[randint(0, 3)])
 
-        if message.content == 'eep' or message.content == 'sleep' or message.content == 'sleepy':
-            await message.channel.send("i eep <@" + str(awex) + "> " + messages[randint(0, 3)])
+            if message.content == 'eep' or message.content == 'sleep' or message.content == 'sleepy':
+                await message.channel.send("i eep <@" + str(awex) + "> " + messages[randint(0, 3)])
 
-        # pings awex in rate song channel when a sentence with the word rate is used
-        if(message.channel.id == 1115319072314372177 and 'rate' in message.content):
+            # pings awex in rate song channel when a sentence with the word rate is used
+            if(message.channel.id == 1115319072314372177 and 'rate' in message.content):
+                await message.channel.send("<@" + str(awex) + "> " + messages[randint(0, 3)])
+
+            # sheryel real preventer
+            if(message.content in real_spellings):
+                await message.delete()
+                await message.channel.send("# stop playing real <@" + str(sheryel) + ">")
+
+            # says hi when stuff
+            if((message.content.lower()).strip('!@#$%^&*(){}|:"<>?\'1234567890-=[]\;,./') in ['hi', 'hello', 'hellp', 'hii']):
+                await message.reply('hi')
+
+            # says something in messages when cherry is mentioned
+            if((message.content.lower()).strip('!@#$%^&*(){}|:"<>?\'1234567890-=[]\;,./') == 'cherry'):
+                await message.reply(messages[randint(0, 3)])
+
+            # your mom
+            if(message.content == 'your mom'):
+                await message.reply('insane burn')
+        
+        if message.author.id == 1114730963600150528 and message.content == 'no' and message.channel.id == 1115319072314372177:
+            await message.delete()
             await message.channel.send("<@" + str(awex) + "> " + messages[randint(0, 3)])
 
-        # sheryel real preventer
-        if(message.content in real_spellings):
-            await message.delete()
-            await message.channel.send("# stop playing real <@" + str(sheryel) + ">")
+        if message.content == "🍄":
+            await message.reply('mushroom')
+        
 
-        # says hi when stuff
-        if((message.content.lower()).strip('!@#$%^&*(){}|:"<>?\'1234567890-=[]\;,./') in ['hi', 'hello', 'hellp', 'hii']):
-            await message.reply('hi')
-
-        # says something in messages when cherry is mentioned
-        if((message.content.lower()).strip('!@#$%^&*(){}|:"<>?\'1234567890-=[]\;,./') == 'cherry'):
-            await message.reply(messages[randint(0, 3)])
-
-        # your mom
-        if(message.content == 'your mom'):
-            await message.reply('insane burn')
-    
-    if message.author.id == 1114730963600150528 and message.content == 'no' and message.channel.id == 1115319072314372177:
-        await message.delete()
-        await message.channel.send("<@" + str(awex) + "> " + messages[randint(0, 3)])
-
-    if message.content == "🍄":
-        await message.reply('mushroom +1')
-        global mushroomnum
-        mushroomnum += 1
-        mushdatafile = open(datapath, 'w')
-        mushdatafile.truncate(0)
-        mushdatafile.write(str(mushroomnum))
-        mushdatafile.close()
         
 
 bot.run(__THING__)
